requirements.txt
alpha-vantage==2.1.0
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
idna==2.8
isort==4.3.20
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
matplotlib==3.1.0
mccabe==0.6.1
numpy==1.16.4
pandas==0.24.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.1
zope.interface==4.6.0
